Assistant checkpoint: Fix holdPeriod validation error
Assistant generated file changes:
- Replit Commercial RE Calc/backend/app/main_simple.py: Fix holdPeriod to accept float values

---

User prompt:

0c6ff1-d267-43d8-8c9f-bf60318d83e9-00-gfnmz72ytcwa.worf.replit.dev says Error analyzing deal: Failed to analyze deal: 422 - {"detail":[{"type":"int_from_float","loc":["body","exitAssumptions","holdPeriod"],"msg":"Input should be a valid integer, got a number with a fractional part","input":6.5}]}

+++ Replit Commercial RE Calc/backend/app/main_simple.py
@@ -32,7 +32,7 @@ class LoanTerms(BaseModel):
     interestOnlyMonths: int = 0
 
 class ExitAssumptions(BaseModel):
-    holdPeriod: int = 5
+    holdPeriod: float = 5.0
     exitCapRate: float = 6.5
     annualAppreciation: float = 3.0
     marketCapRate: float = 6.0
